Add 'lists.md' and update reference in README
A new file, 'lists.md', has been added to provide instructions on how to run and work with lists. Additionally, the link in the README file has been updated to point to this new file instead of directly to the 'lists.py' file. This change provides better guidance for users to follow.